# tsdb: re-use iterator when stepping through chunks